generalize locked progress component
