+++ src/components/locked-progress/index.ts
@@ -1,0 +1,1 @@
+export { LockedProgress } from "./locked-progress";
