+++ src/components/locked-progress/locked-progress.scss
@@ -1,0 +1,36 @@
+@import "../../styles/colors";
+@import "../../styles/fonts";
+
+.tranche-item {
+  &__progress {
+    border-left: $white;
+    border-left-style: solid;
+    border-left-width: 1px;
+    border-right: $white;
+    border-right-style: solid;
+    border-right-width: 1px;
+
+    &-contents {
+      display: flex;
+      justify-content: space-between;
+      gap: 0 5px;
+      font-family: $font-mono;
+      padding: 2px 5px 2px 5px;
+      color: $text-color-deemphasise;
+    }
+  }
+
+  &__progress-bar {
+    display: flex;
+
+    &--locked {
+      background-color: $vega-pink;
+      height: 16px;
+    }
+
+    &--unlocked {
+      background-color: $vega-green;
+      height: 16px;
+    }
+  }
+}
